CER_LoadProfiles/requirements.txt
rampdemand>=0.5.0
pyloadprofilegenerator>=0.1.0
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0
pyyaml>=6.0
matplotlib>=3.7.0
